requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-editorjs==0.2.1
django-filter==23.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==18.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
validate-email==1.3
